:building_construction: minor updates
