+++ .github/workflows/build.yml
@@ -10,7 +10,7 @@ jobs:
   cargo-fmt:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
     - name: Get latest version of stable rust
       run: rustup update stable
     - name: Check formatting with cargofmt
@@ -18,7 +18,7 @@ jobs:
   cargo-clippy:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
     - name: Get latest version of stable rust
       run: rustup update stable
     - name: Check for lint warnings
@@ -27,7 +27,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: cargo-fmt
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Get latest version of stable rust
         run: rustup update stable
       - name: Run tests in release
@@ -36,7 +36,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: cargo-fmt
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Get latest version of stable rust
         run: rustup update stable
       - name: Run tests in release
@@ -45,7 +45,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: cargo-fmt
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Get latest version of stable rust
         run: rustup update stable
       - name: Run tests in release
@@ -54,7 +54,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: cargo-fmt
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Get latest version of stable rust
         run: rustup update stable
       - name: Get latest cargo audit
@@ -67,6 +67,6 @@ jobs:
     container:
       image: rust
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
     - name: Check rustdoc links
       run: RUSTDOCFLAGS="--deny broken_intra_doc_links" cargo doc --verbose --workspace --no-deps --document-private-items
